requirements.txt
absl-py==1.4.0
accelerate==0.15.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cfgv==3.3.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cmake==3.26.0
comm==0.1.3
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
descartes==1.1.0
diffusers==0.14.0
diffuzers==0.3.5
distlib==0.3.6
einops==0.6.0
entrypoints==0.4
executing==1.2.0
facexlib==0.2.5
fairscale==0.4.4
fastapi==0.88.0
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.10.0
filterpy==1.4.5
fire==0.5.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.3
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==1.0.0
gradio==3.23.0
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
identify==2.5.21
idna @ file:///croot/idna_1666125576474/work
imageio==2.26.1
importlib-metadata==6.1.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
kiwisolver==1.4.4
kornia==0.6.11
langcodes==3.3.0
lazy_loader==0.1
linkify-it-py==2.0.0
lit==16.0.0
llvmlite==0.39.1
lmdb==1.4.0
loguru==0.6.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nodeenv==1.7.0
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
nuscenes-devkit==1.1.10
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.9.1
openai==0.27.2
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.8.8
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker==2.7.0
pre-commit==3.2.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.6
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.6
pydeck==0.8.0
pydub==0.25.1
Pygments==2.14.0
Pympler==1.0.1
pyngrok==5.2.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.5
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
realesrgan==0.3.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
salesforce-lavis==1.0.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
Shapely==1.8.5
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-clickable-images==0.0.3
stack-data==0.6.2
starlette==0.22.0
streamlit==1.16.0
streamlit-drawable-canvas==0.9.0
sympy==1.11.1
tb-nightly==2.13.0a20230319
tenacity==8.2.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@cae78c46d658a8e496a815c2ee49b9b178fb9c9a
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uc-micro-py==1.0.1
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
uvicorn==0.15.0
validators==0.20.0
virtualenv==20.21.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.7
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
